-removed releases folder -renamed messageControl refs to colorControls -renamed component.tsx files to match their component names -removed index.tsx files
