+++ src/components/header/__tests__/test_header.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { Header } from "../component";
+import { Header } from "../Header";
 import renderer from "react-test-renderer";
 
 it("component renders", () => {
