+++ src/components/homeButton/__tests__/test_homeButton.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { HomeButton } from "../component";
+import { HomeButton } from "../HomeButton";
 import renderer from "react-test-renderer";
 
 it("component renders", () => {
