--- src/components/homeButton/index.ts
@@ -1,1 +1,0 @@
-export { HomeButton } from "./component";
