--- src/components/messageFormControl/index.tsx
@@ -1,4 +1,0 @@
-export {
-    MessageFormControl as MessageFormControl,
-    MessageFormControlProps,
-} from "./component";
