--- src/components/textFormControls/index.tsx
@@ -1,1 +1,0 @@
-export { TextFormControls, TextFormControlsProps } from "./component";
