--- src/pages/textEditor/index.ts
@@ -1,1 +1,0 @@
-export { TextEditor } from "./component";
