--- src/pages/textEditor/styles.module.css
@@ -1,48 +1,0 @@
-.btn {
-    @apply inline-flex;
-    @apply items-center;
-    @apply justify-center;
-    @apply px-4;
-    @apply py-2;
-    @apply border;
-    @apply border-transparent;
-    @apply text-sm;
-    @apply font-medium;
-    @apply whitespace-nowrap;
-    @apply rounded-md;
-    @apply shadow-sm;
-    @apply text-white;
-    @apply bg-green-600;
-    @apply hover:bg-green-700;
-    @apply focus:outline-none;
-    @apply focus:ring-2;
-    @apply focus:ring-offset-2;
-    @apply focus:ring-green-500;
-    @apply disabled:bg-green-600/0;
-    @apply disabled:text-transparent;
-    @apply duration-300;
-}
-.btnRed {
-    @apply inline-flex;
-    @apply items-center;
-    @apply justify-center;
-    @apply px-4;
-    @apply py-2;
-    @apply border;
-    @apply border-transparent;
-    @apply text-sm;
-    @apply font-medium;
-    @apply whitespace-nowrap;
-    @apply rounded-md;
-    @apply shadow-sm;
-    @apply text-white;
-    @apply bg-red-600;
-    @apply hover:bg-red-700;
-    @apply focus:outline-none;
-    @apply focus:ring-2;
-    @apply focus:ring-offset-2;
-    @apply focus:ring-red-500;
-    @apply disabled:bg-red-600/0;
-    @apply disabled:text-transparent;
-    @apply duration-300;
-}
